Move and refactor RestControllerTest tests to integration test
